Fix 'KeyError: file' in browser tests on Python 3.11.
Fixes part of https://github.com/plone/Products.CMFPlone/issues/3663

See https://github.com/plone/Products.CMFPlone/pull/3673 for a similar fix I did in CMFPlone.
In that case I chose to rewrite the doctest as a unit test.
The main fix in both cases is setting `browser.raiseHttpErrors = False`.
This means that you get the original `zException.NotFound` instead of an HTTPError.
The HTTPError is the one that gives problems in Python 3.11 because too many things have changed in exception handling, mostly for improved display of the traceback.